Purged Inheritance
Inheritance was originally put in the code to allow for multiple versions for different components to coexist easily. As we further developed the code however, this supposed advantage became more of a burden as we had to constantly maintain the abstract classes. Furthermore, we don't have multiple versions of any of our code parts, so it was a useless abstraction.